\name{EarthVSOP87B}
\alias{EarthVSOP87B}
\docType{data}
\title{
Computation of Heliocentric Spherical Longitude and Latitude for the Earth
}
\description{
The data contains coefficients used in computation of heliocentric spherical longitude and latitude for the Earth, with reference to the standard equinox J2000.0 (VSOP87 Series B).
}
\usage{EarthVSOP87B}
\format{
  A data frame with the following four variables.
  \describe{
    \item{\code{A}}{numeric Amplitude}
    \item{\code{B}}{numeric Phase}
    \item{\code{C}}{numeric Frequency}
    \item{\code{Subseries}}{factor VSOP87B L and B sub-series}
    
}
}
\source{
\url{ftp://ftp.imcce.fr/pub/ephem/planets/vsop87/VSOP87B.ear}
}
