\name{magn97}
\alias{magn97}
\docType{data}
\title{
Magnitude data for the year 1997
}
\description{
Visual meteor magnitude dataset for the year 1997.
}
\usage{magn97}
\format{
  A data frame with 9208 observations on the following 29 variables.
  \describe{
    \item{\code{IMOcode}}{factor  IMO observer code}
    \item{\code{sitecode}}{numeric  IMO site code}
    \item{\code{long}}{numeric  Longitude of the observing site in degrees}
    \item{\code{EW}}{factor  East (E) or west (W) position from the prime meridian}
    \item{\code{lat}}{numeric  Latitude of the observing site in degrees}
    \item{\code{NS}}{factor  North (N) or south (S) position from the equator}
    \item{\code{day}}{numeric  Day of the month}
    \item{\code{month}}{numeric  Month of the year}
    \item{\code{year}}{numeric  Year 1997}
    \item{\code{start}}{numeric  Beginning of the observing time period}
    \item{\code{stop}}{numeric  End of the observing time period}
    \item{\code{sollong}}{numeric  Solar longitude of the middle of observing time period}
    \item{\code{Shw}}{factor  Abbreviation of the observed shower}
    \item{\code{lmg}}{numeric  Limiting magnitude}
    \item{\code{m6}}{numeric  Number of observed meteors of -6 magnitude}  
    \item{\code{m5}}{numeric  Number of observed meteors of -5 magnitude}
    \item{\code{m4}}{numeric  Number of observed meteors of -4 magnitude}
    \item{\code{m3}}{numeric  Number of observed meteors of -3 magnitude}
    \item{\code{m2}}{numeric  Number of observed meteors of -2 magnitude}
    \item{\code{m1}}{numeric  Number of observed meteors of -1 magnitude}
    \item{\code{zero}}{numeric  Number of observed meteors of 0 magnitude}
    \item{\code{p1}}{numeric  Number of observed meteors of +1 magnitude}
    \item{\code{p2}}{numeric  Number of observed meteors of +2 magnitude}
    \item{\code{p3}}{numeric  Number of observed meteors of +3 magnitude}
    \item{\code{p4}}{numeric  Number of observed meteors of +4 magnitude}
    \item{\code{p5}}{numeric  Number of observed meteors of +5 magnitude}
    \item{\code{p6}}{numeric  Number of observed meteors of +6 magnitude}
    \item{\code{p7}}{numeric  Number of observed meteors of +7 magnitude}
    \item{\code{N}}{numeric  Total number of observed meteors}
  }
}
\source{
Visual Meteor Database, \url{http://www.imo.net/observations/methods/visual-observation/}
}
