\name{MetFns-package}
\alias{MetFns-package}
\alias{MetFns}
\docType{package}
\title{
Analysis of Visual Meteor Data
}
\description{
Visual meteor data, functions for selection of the data, calculations of 
Zenithal Hourly Rate (ZHR), solar longitude, population index and graphical 
representation of ZHR and magnitude distribution
}
\details{
\tabular{ll}{
Package: \tab MetFns\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-16\cr
License: \tab GPL-2 | GPL-3\cr
}
}
\author{
Kristina Veljkovic

Maintainer: Kristina Veljkovic <mackikac@gmail.com>
}


