\name{popind.err}
\alias{popind.err}
\docType{data}
\title{
Error margins for population index
}
\description{
The data represents error margins for the population index with a given number of meteors and value
of the population index.
}
\usage{popind.err}
\format{
  A data frame with the following three numeric variables.
  \describe{
    \item{\code{r}}{Population index}
    \item{\code{n}}{Number of meteors}
    \item{\code{r.err}}{Error margins for the population index}
    
}
}
\source{
Arlt, R. (2003). Bulletin 19 of the International Leonid Watch: Population index study of the 2002 Leonid meteors. \emph{WGN, Journal of the IMO},31:3, 77-87.
}
