\name{popind}
\alias{popind}
\docType{data}
\title{
Conversion for the population index 
}
\description{
The data represents conversion  from the average distance from the limiting magnitude to the population index.
}
\usage{popind}
\format{
  A data frame with the following two numeric variables.
  \describe{
    \item{\code{r}}{Population index}
    \item{\code{avdeltam}}{Average distance from the limiting magnitude}
    
}
}
\source{
Arlt, R. (2003). Bulletin 19 of the International Leonid Watch: Population index study of the 2002 Leonid meteors. \emph{WGN, Journal of the IMO},31:3, 77-87.
}
