\name{rate11}
\alias{rate11}
\docType{data}
\title{
Rate data for the year 2011
}
\description{
Visual meteor rate data for the year 2011.
}
\usage{rate11}
\format{
  A data frame with 4519 observations on the following 30 variables.
  \describe{
\item{\code{IMOcode}}{factor  IMO observer code}
    \item{\code{sitecode}}{numeric  IMO site code}
    \item{\code{long}}{numeric  Longitude of the observing site in degrees}
    \item{\code{EW}}{factor  East (E) or west (W) position from the prime meridian}
    \item{\code{lat}}{numeric  Latitude of the observing site in degrees}
    \item{\code{NS}}{factor  North (N) or south (S) position from the equator}
    \item{\code{day}}{numeric  Day of the month}
    \item{\code{month}}{numeric  Month of the year}
    \item{\code{year}}{numeric  Year 2011}
    \item{\code{start}}{numeric  Beginning of the observing time period}
    \item{\code{stop}}{numeric  End of the observing time period}
    \item{\code{sollong}}{numeric  Solar longitude of the middle of observing time period}
    \item{\code{fovRA}}{numeric  Right ascension of the center of the field of view}
    \item{\code{fovDEC}}{numeric  Declination of the center of the field of view}
    \item{\code{Teff}}{numeric  Effective observing time}
    \item{\code{F}}{numeric  Correction factor for clouds}
    \item{\code{lmg}}{numeric  Limiting magnitude}
    \item{\code{SPO}}{numeric  Number of observed sporadics}
    \item{\code{Shw1}}{factor  Abbreviation of the first shower} 
    \item{\code{N1}}{numeric  Number of meteors belonging to the first shower}
    \item{\code{Shw2}}{factor  Abbreviation of the second shower} 
    \item{\code{N2}}{numeric  Number of meteors belonging to the second shower}
    \item{\code{Shw3}}{factor  Abbreviation of the third shower} 
    \item{\code{N3}}{numeric  Number of meteors belonging to the third shower}
    \item{\code{Shw4}}{factor  Abbreviation of the forth shower}
    \item{\code{N4}}{numeric  Number of meteors belonging to the forth shower}
    \item{\code{Shw5}}{factor  Abbreviation of the fifth shower}
    \item{\code{N5}}{numeric  Number of meteors belonging to the fifth shower}
    \item{\code{Shw6}}{factor  Abbreviation of the 6th shower}
    \item{\code{N6}}{numeric  Number of meteors belonging to the 6th shower}
  }
}
\source{
Visual Meteor Database, \url{http://www.imo.net/data/visual}
}

