\name{vmdbpers}
\alias{vmdbpers}
\docType{data}
\title{
List of observers
}
\description{
The data consists of a list of observers.
}
\usage{vmdbpers}
\format{
  A data frame with 5920 observations on the following 4 factor variables.
  \describe{
    \item{\code{Obsvr}}{IMO observer code}
    \item{\code{Name}}{Observer's last name}
    \item{\code{Firstname}}{Observer's first name}
    \item{\code{Country}}{Observer's country}
}
}
\source{
Visual Meteor Database, \url{http://www.imo.net/data/visual}
}
