\name{MetFns-package}
\alias{MetFns-package}
\alias{MetFns}
\docType{package}
\title{
Analysis of Visual Meteor Data
}
\description{
Functions for selection of visual meteor data, calculations of Zenithal Hourly 
Rate (ZHR) and population index, graphics of population index, ZHR and 
magnitude distribution.
}
\details{
\tabular{ll}{
Package: \tab MetFns\cr
Type: \tab Package\cr
Version: \tab 3.0.0\cr
Date: \tab 2017-08-18\cr
License: \tab GPL-2 | GPL-3\cr
}
}
\author{
Kristina Veljkovic

Maintainer: Kristina Veljkovic <mackikac@gmail.com>
}


